Hardened and add tests for Server
The Server was prone to error, untested and closed the connection after
receiving 1 message. This has all been fixed in this commit.

Processing a message on a connection is split up into a few function
calls which are well tested. All calls to panic() are removed.

The Server now has ErrorLog attribute so one can use hook up a Server to
his favorite logging engine.

Issue: #7