Add Value type to support unsigned values
Modbus supports 16 bits values. Depending on how one interpret a 16 bit
value, a 16 bit value can represent all values in range of -32768
through 65535.

Goldfish used int16 for values. The problem is that int16 only can take
values in range of -32768 through 32767. Making it not possible to work
in the handlers with the values in range 32768 through 65535.

To solve this a new Value type has been introduced. It's can hold all
values in range of -32768 through 65535.

The signature of goldfish.NewReadHandler() and
goldfish.NewWriteHandler() has been changed to reflects these changes.

Issues: #10