Fix EOF errors when reading requests.
Goldfish printed a lot of EOF errors when reading multiple requests from a single connection. When Goldfish read beyond the last bit of data on a connection Go returned an EOF error. This is not an error in the traditional sense, but an indication the connection is done writing (for now). The connection can be closed after an EOF error, but its not neccesary to log it as it is not a fault in the program.