This commit  fixes timeout issues. (#23)
* This commit  fixes timeout issues.

The server did not implement any sort of timeout. Therefore connections
could stay alive forever, while new connections get added all the time.
This would cause crashes in applications using goldfish.

This commit adds an optional, configurable timeout to the server. This
is done in such a way that the API is fully backwards compatible.